requirements.txt
google-generativeai
python-dotenv
pandas
requests